fixed headers, refactoring
