+++ CMakeLists.txt
@@ -1,6 +1,6 @@
 cmake_minimum_required (VERSION 3.7)
 project (mastodon-cpp
-         VERSION 0.7.19
+         VERSION 0.7.20
          LANGUAGES CXX
 )
