+++ src/easy/context.hpp
@@ -23,12 +23,12 @@
 // If we are compiling mastodon-cpp, use another include path
 #ifdef MASTODON_CPP
     #include "mastodon-cpp.hpp"
-    #include "easy.hpp"
-    #include "status.hpp"
+    #include "easy/easy.hpp"
+    #include "easy/status.hpp"
 #else
     #include <mastodon-cpp/mastodon-cpp.hpp>
-    #include <mastodon-cpp/easy.hpp>
-    #include <mastodon-cpp/status.hpp>
+    #include <mastodon-cpp/easy/easy.hpp>
+    #include <mastodon-cpp/easy/status.hpp>
 #endif
 
 using std::string;
