+++ src/easy/easy.hpp
@@ -136,9 +136,6 @@ public:
     static const std::vector<stream_event>
         parse_stream(const std::string &streamdata);
 
-    /*!
-     *  @brief  Base class for entities.
-     */
     class Entity
     {
     public:
