+++ src/easy/instance.hpp
@@ -23,11 +23,11 @@
 // If we are compiling mastodon-cpp, use another include path
 #ifdef MASTODON_CPP
     #include "mastodon-cpp.hpp"
-    #include "easy.hpp"
-    #include "account.hpp"
+    #include "easy/easy.hpp"
+    #include "easy/account.hpp"
 #else
     #include <mastodon-cpp/mastodon-cpp.hpp>
-    #include <mastodon-cpp/easy.hpp>
+    #include <mastodon-cpp/easy/easy.hpp>
     #include <mastodon-cpp/easy/account.hpp>
 #endif
